refactor: move Errors enum to validator package
