+++ src/main/java/lotto/validator/Errors.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.validator;
 
 import lotto.util.Constants;
 import lotto.util.InputParser;
